+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 ### Added
 
 - `setup` key for tasks to refine service time modeling (#358)
-- `max_tasks` key limiting route size at vehicle level (#421)
+- `max_tasks` key limiting route size at vehicle level (#421, #566)
 - Support for custom cost matrices (#415)
+- number of routes in solution summary (#524)
 - Github Actions CI (#436)
 - Check for libvroom example build in CI (#514)